feat: add pod members
